requirements.txt
Flask-Principal==0.4.0
Flask==2.0.0
Pygments==2.9.0
Werkzeug==2.0.1
bleach==3.3.0
feedgen==0.9.0
httpx==0.18.1
jsonpatch==1.32
marshmallow==3.12.1
pytest==6.2.4
pytest-datadir==1.3.1
rdflib-jsonld==0.5.0
requests==2.25.1
webargs==8.0.0
